remove cloudfront; remove featured_pages_block override
CloudFront requires more work to be released to production.  This reverts the change in 2.17.9.

The override of featured_pages_block was attempting to address raising of an exception if a deleted page was referrenced.  The override didn’t work.  It removed all pages.  A fix has since been made in spotlight that does work.