Remove the Video widget
Uploading videos through the video widget generally does not work and leaves the page in a corrupt state requiring manual intervention to repair.  See Issue #147 which describes the manual process and requests removal of the video widget.

NOTE: Documentation needs to be added to describe the alternate approach using Embed + Text or Iframe Embed.