release 2.19.0
revert removal of video widget

Without the video widget enabled, editing a page that already has a video caused the video to be removed.  Our goal was to prevent upload of new videos through the video widget.

Have to revert for now to come up with a better long term approach.