add warning messages for duplicate files
Files live under `.platform/hooks` and `.platform/confighooks`.  The first runs when deployed and the second runs when there are configuration changes.  The same scripts are needed for both events and so are duplicated.  The warning is an attempt to avoid changes happening to a file under one directory, but not the other.